Added ScoreAuditor.  Changed probablity in Auditor to a Float.

+++ aloha-core/src/test/scala/com/eharmony/aloha/score/audit/take5/Auditor.scala
@@ -12,6 +12,15 @@ import com.eharmony.aloha.reflect.RefInfo
 trait Auditor[U, -N, +B <: U] {
   type OutputType[+X] <: U
 
+  /**
+    * Should this be done via implicit resolution?
+    * {{{
+    * def changeType[M](implicit auditor: Auditor[U, M, OutputType[M]]): Auditor[U, M, OutputType[M]] = auditor
+    * }}}
+    * @tparam M a new natural type.
+    * @return
+    */
+  // TODO: Figure out if we should do this via implicit score resolution.
   def changeType[M: RefInfo]: Option[Auditor[U, M, OutputType[M]]]
 
   private[aloha] def failure(key: ModelIdentity,
@@ -23,6 +32,6 @@ trait Auditor[U, -N, +B <: U] {
                              valueToAudit: N,
                              missingVarNames: => Set[String],
                              subValues: Seq[U],
-                             prob: => Option[Double]): B
+                             prob: => Option[Float]): B
 }
